+++ packages/react/test/spec/Components/Data/WebtrekkAdvancedData.test.js
@@ -26,6 +26,14 @@ describe('WebtrekkAdvancedData', () => {
                     requestObfuscation={ 'foo' }
                     execCDB={ 1 }
                     useCDBCache={ 1 }
+                    sendViaSDK={ 1 }
+                    sendViaServer={ {
+                        activated: 0,
+                        serverDomain: false,
+                        serverPath: false,
+                        droppedRequests: '',
+                        blacklist: ''
+                    } }
                     useHashForDefaultPageName={ 1 }
                     useParamsForDefaultPageName={ 'param' }
                     requestQueue={ {
@@ -38,16 +46,18 @@ describe('WebtrekkAdvancedData', () => {
             );
 
             expect(spyOnError).toHaveBeenCalled();
-            expect(spyOnError.mock.calls.length).toBe(8);
+            expect(spyOnError.mock.calls.length).toBe(10);
 
             expect(spyOnError.mock.calls[0][0]).toContain('Invalid prop `secureCookie` of type `string` supplied to `WebtrekkAdvancedData`, expected `boolean`.');
             expect(spyOnError.mock.calls[1][0]).toContain('Invalid prop `optOutName` of type `number` supplied to `WebtrekkAdvancedData`, expected `string`.');
             expect(spyOnError.mock.calls[2][0]).toContain('Invalid prop `requestObfuscation` of type `string` supplied to `WebtrekkAdvancedData`, expected `boolean`.');
             expect(spyOnError.mock.calls[3][0]).toContain('Invalid prop `execCDB` of type `number` supplied to `WebtrekkAdvancedData`, expected `boolean`.');
             expect(spyOnError.mock.calls[4][0]).toContain('Invalid prop `useCDBCache` of type `number` supplied to `WebtrekkAdvancedData`, expected `boolean`.');
-            expect(spyOnError.mock.calls[5][0]).toContain('Invalid prop `useHashForDefaultPageName` of type `number` supplied to `WebtrekkAdvancedData`, expected `boolean`.');
-            expect(spyOnError.mock.calls[6][0]).toContain('Invalid prop `useParamsForDefaultPageName` of type `string` supplied to `WebtrekkAdvancedData`, expected an array.');
-            expect(spyOnError.mock.calls[7][0]).toContain('Invalid prop `requestQueue.activated` of type `number` supplied to `WebtrekkAdvancedData`, expected `boolean`.');
+            expect(spyOnError.mock.calls[5][0]).toContain('Invalid prop `sendViaSDK` of type `number` supplied to `WebtrekkAdvancedData`, expected `boolean`.');
+            expect(spyOnError.mock.calls[6][0]).toContain('Invalid prop `sendViaServer.activated` of type `number` supplied to `WebtrekkAdvancedData`, expected `boolean`.');
+            expect(spyOnError.mock.calls[7][0]).toContain('Invalid prop `useHashForDefaultPageName` of type `number` supplied to `WebtrekkAdvancedData`, expected `boolean`.');
+            expect(spyOnError.mock.calls[8][0]).toContain('Invalid prop `useParamsForDefaultPageName` of type `string` supplied to `WebtrekkAdvancedData`, expected an array.');
+            expect(spyOnError.mock.calls[9][0]).toContain('Invalid prop `requestQueue.activated` of type `number` supplied to `WebtrekkAdvancedData`, expected `boolean`.');
         });
 
         test('don\'t returns children', () => {
@@ -78,10 +88,18 @@ describe('WebtrekkAdvancedData', () => {
         test('mount advanced', (done) => {
             mount(<WebtrekkAdvancedData
                 secureCookie={ true }
-                optOutName='webtrekkTestOptOut'
+                optOutName="webtrekkTestOptOut"
                 requestObfuscation={ true }
                 execCDB={ false }
                 useCDBCache={ true }
+                sendViaSDK={ true }
+                sendViaServer={ {
+                    activated: true,
+                    serverDomain: 'sub.domain.tld',
+                    serverPath: 'path/to/s2s',
+                    droppedRequests: 0,
+                    blacklist: ['foo', 'bar']
+                } }
                 useHashForDefaultPageName={ true }
                 useParamsForDefaultPageName={ ['param1', 'param2'] }
                 requestQueue={ {
@@ -100,6 +118,12 @@ describe('WebtrekkAdvancedData', () => {
                     expect(advancedData.requestObfuscation).toBe(true);
                     expect(advancedData.execCDB).toBe(false);
                     expect(advancedData.useCDBCache).toBe(true);
+                    expect(advancedData.sendViaSDK).toBe(true);
+                    expect(advancedData.sendViaServer.activated).toBe(true);
+                    expect(advancedData.sendViaServer.serverDomain).toBe('sub.domain.tld');
+                    expect(advancedData.sendViaServer.serverPath).toBe('path/to/s2s');
+                    expect(advancedData.sendViaServer.droppedRequests).toBe(0);
+                    expect(advancedData.sendViaServer.blacklist).toEqual(['foo', 'bar']);
                     expect(advancedData.useHashForDefaultPageName).toBe(true);
                     expect(advancedData.useParamsForDefaultPageName.join(',')).toBe('param1,param2');
                     expect(advancedData.requestQueue.activated).toBe(true);
@@ -113,10 +137,18 @@ describe('WebtrekkAdvancedData', () => {
         test('update advanced', (done) => {
             renderedWebtrekkAdvancedData = mount(<WebtrekkAdvancedData
                 secureCookie={ true }
-                optOutName='webtrekkTestOptOut'
+                optOutName="webtrekkTestOptOut"
                 requestObfuscation={ true }
                 execCDB={ false }
                 useCDBCache={ true }
+                sendViaSDK={ true }
+                sendViaServer={ {
+                    activated: true,
+                    serverDomain: 'sub.domain.tld',
+                    serverPath: 'path/to/s2s',
+                    droppedRequests: 0,
+                    blacklist: ['foo', 'bar']
+                } }
                 useHashForDefaultPageName={ true }
                 useParamsForDefaultPageName={ ['param1', 'param2'] }
                 requestQueue={ {
@@ -141,6 +173,12 @@ describe('WebtrekkAdvancedData', () => {
                     expect(advancedData.requestObfuscation).toBe(true);
                     expect(advancedData.execCDB).toBe(true);
                     expect(advancedData.useCDBCache).toBe(false);
+                    expect(advancedData.sendViaSDK).toBe(true);
+                    expect(advancedData.sendViaServer.activated).toBe(true);
+                    expect(advancedData.sendViaServer.serverDomain).toBe('sub.domain.tld');
+                    expect(advancedData.sendViaServer.serverPath).toBe('path/to/s2s');
+                    expect(advancedData.sendViaServer.droppedRequests).toBe(0);
+                    expect(advancedData.sendViaServer.blacklist).toEqual(['foo', 'bar']);
                     expect(advancedData.useHashForDefaultPageName).toBe(true);
                     expect(advancedData.useParamsForDefaultPageName.join(',')).toBe('param1,param2');
                     expect(advancedData.requestQueue.activated).toBe(true);